CADC-11734 refactored ResourceStats and GetAction in sessions based on Brian's code review. Added units to RAM output.
